Add new variant of location request method with delayUntilAuthorized parameter
Merge pull request #2 from @podkovyrin.

By default, the timeout countdown begins as soon as the `requestLocationWithDesiredAccuracy:timeout:block:` method is called. This commit introduces a new variant of this method that includes a `delayUntilAuthorized` parameter, which allows you to pass YES to delay the start of the timeout countdown until the user has responded to the system location services permissions prompt (if the user hasn't allowed or denied the app access yet).