Fix statusForLocationRequest: to correctly return Disabled
Merge PR #27 from @KhanFu

As the code was originally written, the check for kCLAuthorizationStatusDenied came first and thus would always short-circuit because that authorization status is returned when "User has explicitly denied authorization for this application, or location services are disabled in Settings." (per Apple's docs).

Moving the check of +[CLLocationManager locationServicesEnabled] ahead of it correctly detects the case where location services are disabled device-wide (even if this app does have permission to use them).