Make locationRequests property type NSArray (immutable)
Instead of keeping a mutable array of location requests, make this property type NSArray. When we need to mutate the array (adding or removing a location request), this is done by making a mutable copy of the existing array, mutating that copy, then assigning the new array to the property. This resolves any issues around mutating the location requests array while iterating/enumerating over it, as it is no longer possible to directly mutate the array, and fast enumeration will use the same instance even if additional "versions" (instances) of the array are set to the property during enumeration. This also cleans up existing code where we already were using a separate mutable array to avoid mutating while enumerating.

Fixes #46