Support background location updates in iOS 9
iOS 9 requires setting allowsBackgroundLocationUpdates to YES in order to receive background location updates. We only set it to YES if the location background mode is enabled for this app, as the documentation suggests it is a fatal programmer error otherwise. Additionally enabled the location background mode for the example app.